refactor(useAreaOutsideClick): useOuterClickを実体に併せてuseAreaOutsideClickに変更
